API

Public

YeSQL.load_path — Function

load_path(path; glob_pattern) -> Dict{Symbol, YeSQL.SQLQuery}

Read files(s) at path and return extracted SQLQuerys. If path is a directory then include files matching glob_pattern.

YeSQL.load_path! — Function

load_path!(queries::AbstractDict, path; glob_pattern)

Read files(s) at path and insert extracted SQLQuerys into queries. If path is a directory then include files matching glob_pattern.

YeSQL.load_string — Function

load_string(query_text::AbstractString) -> Dict{Symbol, YeSQL.SQLQuery}

Parse query_text and return extracted SQLQuerys

YeSQL.load_string! — Function

load_string!(queries::AbstractDict, query_text::AbstractString)

Parse query_text and insert extracted SQLQuerys into queries

Internals

Query representation

YeSQL.SQLQuery — Type

struct SQLQuery{O<:YeSQL.SQLOperation}

SQL query with fields extracted from a .sql file or string

Fields

name::Symbol

Query name

query::String

Processed query string

docstring::String

Docstring

args::Vector{Symbol}

Argument list

operation::YeSQL.SQLOperation

SQLOperation derived from the operation suffix

Query syntax

Each query should be preceded by comment lines defining the query name and operation type, and an optional docstring:

-- name: <name><operation_suffix>
-- <docstring>  (optional)
<query text>

SQL Operations

These empty types are used for dispatch when executing a query. When queries are loaded the query name suffix is used to determine the appropriate operation.

YeSQL.SQLOperation — Type

Abstract parent type

YeSQL.Select — Type

A select query returning potentially many rows. No suffix.

YeSQL.SelectOne — Type

A query returning a single row. Name suffix: ^

YeSQL.SelectValue — Type

A query returning a single value. Name suffix: $

YeSQL.InsertUpdateDelete — Type

An operation which inserts/updates/deletes a row. Name suffix: !

YeSQL.InsertUpdateDeleteMany — Type

An operation which inserts/updates/deletes many rows. Name suffix: *!

YeSQL.InsertReturning — Type

An operation which inserts a row returns a row of values. Name suffix: <!

YeSQL.Script — Type

A sql script run without arguments. Name suffix: #

Functions

YeSQL._interpret — Function

_interpret(query_name::AbstractString) -> Any

Interpret query_name and return an SQLOperation and name as a symbol

YeSQL._interpret("my_script_query#")

# output

(YeSQL.Script(), :my_script_query)

YeSQL._extract! — Function

_extract!(queries::AbstractDict, queries_text::AbstractString)

Parse queries_text and insert extracted SQLQuerys into queries